ingestor-requirements.txt
psycopg2-binary==2.9.7
PyYAML==6.0.1
requests==2.31.0
